requirements.txt
Django==1.10.5
django-widget-tweaks==1.4.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.7
whitenoise==3.3.0
djangorestframework==3.8.2
markdown==2.6.11
django-filter==1.1.0
fcm_django==0.2.19
coreapi==2.3.3